+++ src/CellsContainer.php
@@ -14,8 +14,8 @@ declare(strict_types=1);
 
 namespace MarcinOrlowski\AsciiTable;
 
-use MarcinOrlowski\AsciiTable\Exceptions\ColumnKeyNotFound;
-use MarcinOrlowski\AsciiTable\Exceptions\DuplicateColumnKey;
+use MarcinOrlowski\AsciiTable\Exceptions\ColumnKeyNotFoundException;
+use MarcinOrlowski\AsciiTable\Exceptions\DuplicateColumnKeyException;
 use MarcinOrlowski\AsciiTable\Traits\ArrayAccessTrait;
 use MarcinOrlowski\AsciiTable\Traits\IteratorAggregateTrait;
 
@@ -33,12 +33,12 @@ class CellsContainer implements ContainerContract
      * @param string|int $columnKey Key of the column we want this cell to belong to.
      * @param Cell       $cell      Instance of `Cell` to be added.
      *
-     * @throws DuplicateColumnKey
+     * @throws DuplicateColumnKeyException
      */
     public function add(string|int $columnKey, Cell $cell): self
     {
         if ($this->offsetExists($columnKey)) {
-            throw new DuplicateColumnKey("Column key already exists: {$columnKey}");
+            throw new DuplicateColumnKeyException("Column key already exists: {$columnKey}");
         }
 
         $this->container[ $columnKey ] = $cell;
@@ -50,12 +50,12 @@ class CellsContainer implements ContainerContract
      *
      * @param string|int $columnKey Key of the column we want this cell to belong to.
      *
-     * @throws ColumnKeyNotFound
+     * @throws ColumnKeyNotFoundException
      */
     public function get(string|int $columnKey): Cell
     {
         if (!$this->offsetExists($columnKey)) {
-            throw new ColumnKeyNotFound("Unknown column key: {$columnKey}");
+            throw new ColumnKeyNotFoundException("Unknown column key: {$columnKey}");
         }
         return $this->container[ $columnKey ];
     }