+++ src/Exceptions/NoVisibleColumnsException.php
@@ -1,0 +1,23 @@
+<?php
+declare(strict_types=1);
+
+/**
+ * ASCII Table
+ *
+ * @package   MarcinOrlowski\AsciiTable
+ *
+ * @author    Marcin Orlowski <mail (#) marcinOrlowski (.) com>
+ * @copyright 2022 Marcin Orlowski
+ * @license   http://www.opensource.org/licenses/mit-license.php MIT
+ * @link      https://github.com/MarcinOrlowski/php-ascii-table
+ */
+
+namespace MarcinOrlowski\AsciiTable\Exceptions;
+
+class NoVisibleColumnsException extends \Exception
+{
+    public function __construct()
+    {
+        parent::__construct('No visible columns in table. Enable some?');
+    }
+}